Move undo/redo functions and types into separate module
