+++ src/client/reducers/districtDrawing.ts
@@ -1,4 +1,4 @@
-import { Cmd, CmdType, Loop, loop, LoopReducer } from "redux-loop";
+import { Cmd, Loop, loop, LoopReducer } from "redux-loop";
 import { getType } from "typesafe-actions";
 
 import { Action } from "../actions";
@@ -29,15 +29,15 @@ import {
 import { updateDistrictsDefinition } from "../actions/projectData";
 import { SelectionTool } from "../actions/districtDrawing";
 import { resetProjectState } from "../actions/root";
-import {
-  DistrictsDefinition,
-  GeoUnits,
-  GeoUnitsForLevel,
-  LockedDistricts
-} from "../../shared/entities";
+import { GeoUnits, GeoUnitsForLevel } from "../../shared/entities";
 import { ProjectState, initialProjectState } from "./project";
-
-const UNDO_HISTORY_MAX_LENGTH = 100;
+import {
+  pushEffect,
+  pushStateUpdate,
+  UndoHistory,
+  UndoableState,
+  updateCurrentState
+} from "./undoRedo";
 
 function setGeoUnitsForLevel(
   currentGeoUnits: GeoUnitsForLevel,
@@ -80,79 +80,6 @@ function clearGeoUnits(geoUnits: GeoUnits): GeoUnits {
       [geoLevelId]: new Map()
     };
   }, {});
-}
-
-function truncatePastUndoHistory(past: readonly UndoableStateAndEffect[]) {
-  // Limit the undo history to the n most recent items
-  return past.slice((UNDO_HISTORY_MAX_LENGTH - 1) * -1);
-}
-
-function pushStateUpdate(state: ProjectState, undoState: Partial<UndoableState>): ProjectState {
-  return {
-    ...state,
-    undoHistory: {
-      past: [...truncatePastUndoHistory(state.undoHistory.past), state.undoHistory.present],
-      present: {
-        state: {
-          ...state.undoHistory.present.state,
-          ...undoState
-        }
-      },
-      future: []
-    }
-  };
-}
-
-function pushEffect(state: ProjectState, effect: Effect): ProjectState {
-  return {
-    ...state,
-    undoHistory: {
-      past: [...truncatePastUndoHistory(state.undoHistory.past), state.undoHistory.present],
-      present: {
-        ...state.undoHistory.present,
-        effect
-      },
-      future: []
-    }
-  };
-}
-
-export function updateCurrentState(state: ProjectState, undoState: Partial<UndoableState>) {
-  return {
-    ...state,
-    undoHistory: {
-      ...state.undoHistory,
-      present: {
-        ...state.undoHistory.present,
-        state: {
-          ...state.undoHistory.present.state,
-          ...undoState
-        }
-      }
-    }
-  };
-}
-
-interface UndoableState {
-  readonly selectedGeounits: GeoUnits;
-  readonly geoLevelIndex: number; // Index is based off of reversed geoLevelHierarchy in static metadata
-  readonly geoLevelVisibility: ReadonlyArray<boolean>; // Visibility values at indices corresponding to `geoLevelIndex`
-  readonly lockedDistricts: LockedDistricts;
-  readonly districtsDefinition: DistrictsDefinition;
-}
-
-// Accompanying effect builder for state update when undone/redone (eg. saving districts definition)
-type Effect = (state: UndoableState) => CmdType<Action>;
-
-interface UndoableStateAndEffect {
-  readonly state: UndoableState;
-  readonly effect?: Effect;
-}
-
-export interface UndoHistory {
-  readonly past: readonly UndoableStateAndEffect[];
-  readonly present: UndoableStateAndEffect;
-  readonly future: readonly UndoableStateAndEffect[];
 }
 
 export interface DistrictDrawingState {
